requirements.txt
Django>=3.2.12,<3.3
psycopg2>=2.8.6,<2.9
django-import-export==2.8.0
requests==2.28.0
djangorestframework==3.13.1